// slider3/types.ts
export type SliderDraggingType = "mouse" | "touch" | null;

/**
 *
 * @internal
 */
export interface SliderState {
  dragging: boolean;
  dragType: SliderDraggingType;
  value: number;
}

/**
 *
 * @internal
 */
export type SliderStepParts = Required<
  Pick<SliderStepOptions, "min" | "max" | "step">
>;

/**
 *
 * @internal
 */
export interface SliderDragValues extends SliderStepParts {
  clientX: number;
  clientY: number;
  top: number;
  height: number;
  left: number;
  width: number;
  vertical: boolean;
  reversed: boolean;
}

export interface SliderIncrementAction extends SliderStepParts {
  type: "increment";
  dragging: boolean;
}

export interface SliderDecrementAction extends SliderStepParts {
  type: "decrement";
  dragging: boolean;
}

export interface SliderMinimumAction {
  type: "minimum";
  min: number;
}

export interface SliderMaximumAction {
  type: "maximum";
  max: number;
}

export interface SliderDragAction extends SliderDragValues {
  type: "drag";
  eventType:
    | "mousedown"
    | "mousemove"
    | "mouseup"
    | "touchstart"
    | "touchmove"
    | "touchend";
}
// ...
export interface SliderRangeOptions {
  /**
   * The min value for the slider.
   */
  min?: number;

  /**
   * The max value for the slider.
   */
  max?: number;

  /**
   * Boolean if the slider is rendered vertically instead of horizontally.
   */
  vertical?: boolean;

  /**
   * Boolean if the slider is disabled and the values cannot be changed.
   */
  disabled?: boolean;
}

export interface SliderStepOptions extends SliderRangeOptions {
  /**
   * A positive number representing the value to "jump" while incrementing or
   * decrementing the slider's value. This should normally stay as the default
   * value of `1`, but can also be decimal values if needed.
   */
  step?: number;
}
// ...
export type SliderEventHandlerNames =
  | "onKeyUp"
  | "onKeyDown"
  | "onMouseDown"
  | "onTouchStart";

export type SliderEventHandlers = Pick<
  HTMLAttributes<HTMLSpanElement>,
  SliderEventHandlerNames
>;

export type OneThumbSliderValue = number;
export type TwoThumbSliderValue = readonly [number, number];
export type SliderValue = OneThumbSliderValue | TwoThumbSliderValue;

export type ReturnedValue<DV extends SliderValue | undefined> = {
  one: OneThumbSliderValue;
  two: TwoThumbSliderValue;
}[DV extends TwoThumbSliderValue ? "two" : "one"];
export type DefaultValue<DV extends SliderValue | undefined> =
  | ReturnedValue<DV>
  | (() => ReturnedValue<DV>);

export interface UseSliderOptions<DV extends SliderValue | undefined>
  extends SliderStepOptions,
    SliderEventHandlers {
  /**
   * An optional ref to merge with the internal `ref` required for the slider to
   * work.
   */
  ref?: Ref<HTMLSpanElement>;

  /**
   * The default value for the slider.
   */
  defaultValue?: DefaultValue<DV>;

  mode?: "onDrag" | "onBlur";
  onValueChange?(value: number): void;
}

export interface SliderBehaviorProps<DV extends SliderValue | undefined>
  extends Required<SliderRangeOptions>,
    Required<SliderEventHandlers> {
  /**
   * A ref that should be passed to the `Slider` component to handle calculating
   * the value based on the drag distance.
   */
  ref: RefCallback<HTMLSpanElement>;

  /**
   * The current value of the slider.
   */
  value: ReturnedValue<DV>;

  dragging: -1 | 0 | 1;
}

// export type UseSliderReturnValue<DV extends SliderValue | undefined> = [
//   SliderBehaviorProps<DV>,
//   SliderActionCreators
// ];
export type UseSliderReturnValue<DV extends SliderValue | undefined> = [
  SliderBehaviorProps<DV>
  // SliderActionCreators
];